Fixed a bug that stopped builds
Using UnityEditorInternal in a script that is not an Editor
folder(TileMain.cs), caused errors when building and stopped the builds.

This is because that namespace isn't supposed to be compiled with
builds, the fix is simple, I just moved the method that used it to the
editor script and made appropriate changes.