Ability to move tiles without breaking system
This update gives the ability to move tiles manually while keeping the
tile add/remove system in one piece.

The system works by redoing how adding/removing tiles works, it now uses
raycasting to check for tiles and edit them accordingly.
Colliders are also added to all tiles in order for raycasting to work,
but these colliders are only temporary for tiles where the user didn't
want colliders and are removed when the game starts.